fix: give self-hosted admins a token form before first connection
(cherry picked from commit a5060646a560140f30b6a1d3297740b5e49c344d)
