+++ packages/dashboard/src/features/webscraper/components/__tests__/WebScraperSettingsDialog.test.tsx
@@ -39,7 +39,12 @@ const connection = {
   createdAt: '2026-03-04T05:06:07.000Z',
 };
 
-function renderDialog() {
+function renderDialog({
+  connection: connectionOverride = connection,
+}: { connection?: typeof connection | null } = {}) {
+  // Default parameters only fall back on `undefined`, not on `null` — unlike
+  // `??`, which would treat an explicitly-passed `connection: null` the same
+  // as "not provided" and silently render the wrong (connected) branch.
   const client = new QueryClient({ defaultOptions: { queries: { retry: false } } });
   return render(
     <QueryClientProvider client={client}>
@@ -47,7 +52,7 @@ function renderDialog() {
         <WebScraperSettingsDialog
           open
           onOpenChange={() => {}}
-          connection={connection}
+          connection={connectionOverride}
           projectId="p1"
         />
       </ToastProvider>
@@ -71,15 +76,10 @@ describe('WebScraperSettingsDialog', () => {
     renderDialog();
 
     await userEvent.click(await screen.findByRole('button', { name: /replace token/i }));
-    await userEvent.type(
-      screen.getByLabelText(/apify api token/i),
-      'apify_api_rotated1234567'
-    );
+    await userEvent.type(screen.getByLabelText(/apify api token/i), 'apify_api_rotated1234567');
     await userEvent.click(screen.getByRole('button', { name: /^save$/i }));
 
-    await waitFor(() =>
-      expect(updateApifyConfig).toHaveBeenCalledWith('apify_api_rotated1234567')
-    );
+    await waitFor(() => expect(updateApifyConfig).toHaveBeenCalledWith('apify_api_rotated1234567'));
   });
 
   it('hides the token controls when the host runs OAuth', async () => {
@@ -99,14 +99,38 @@ describe('WebScraperSettingsDialog', () => {
     renderDialog();
 
     await userEvent.click(await screen.findByRole('button', { name: /replace token/i }));
-    await userEvent.type(
-      screen.getByLabelText(/apify api token/i),
-      'apify_api_rotated1234567'
-    );
+    await userEvent.type(screen.getByLabelText(/apify api token/i), 'apify_api_rotated1234567');
     await userEvent.click(screen.getByRole('button', { name: /^save$/i }));
 
     await waitFor(() => expect(updateApifyConfig).toHaveBeenCalled());
     // The draft stays put after a rejected save — nothing throws, nothing unmounts.
     expect(screen.getByLabelText(/apify api token/i)).toBeInTheDocument();
+  });
+
+  it('shows a working token entry form when self-hosted and not yet connected', async () => {
+    renderDialog({ connection: null });
+
+    // Reachability: WebscraperSidebar's gear button opens this dialog even
+    // when disconnected (disabled only on `!projectId`), so self-hosted
+    // admins land here with `connection === null` before ever configuring a
+    // token. There must be a real way forward, not a dead end: typing a
+    // token must enable the submit button (unlike the old bug, where the
+    // one and only button was permanently disabled because `onConnectApify`
+    // is never defined in self-hosting).
+    const tokenInput = await screen.findByLabelText(/apify api token/i);
+    const connectButton = screen.getByRole('button', { name: /connect apify/i });
+    expect(connectButton).toBeDisabled();
+
+    await userEvent.type(tokenInput, 'apify_api_tok1234567890');
+    expect(connectButton).not.toBeDisabled();
+  });
+
+  it('shows the enabled OAuth button (no token input) when the host runs OAuth and is not yet connected', async () => {
+    hostValue.onConnectApify = vi.fn();
+    renderDialog({ connection: null });
+
+    const connectButton = await screen.findByRole('button', { name: /connect apify/i });
+    expect(connectButton).not.toBeDisabled();
+    expect(screen.queryByLabelText(/apify api token/i)).toBeNull();
   });
 });
